datasworn/pyproject.toml
[project]
name = "pysworn-datasworn"
dynamic = ["version"]
description = "Datasworn Python Wrapper Package"
readme = "README.md"
authors = [{ name = "Gerhard Brandt", email = "gbrandt@cern.ch" }]
# license = { file = "../CC.BY-NC-SA-4.0.LICENSE" }
requires-python = ">=3.12"
dependencies = [
    "cairosvg>=2.8.2",
    "orjson>=3.10.18",
    "rich-pixels>=3.0.1",
    "typer>=0.19.2",
]

[project.scripts]
datasworn = "pysworn.datasworn.main:app"

[tool.hatch.version]
path = 'src/pysworn/core/__init__.py'

[dependency-groups]
dev = ["pytest>=8.4.2", "pytest-asyncio>=1.2.0", "pytest-coverage>=0.0"]
